Avoid compilation warnings:
libfunctions.c: In function ‘load_from_memory’:
libfunctions.c:96: warning: passing argument 1 of ‘gn_cfg_memory_read’ from incompatible pointer type
../include/gnokii.h:271: note: expected ‘const char **’ but argument is of type ‘char **’
cfgreader.c: In function ‘get_locations’:
cfgreader.c:1100: warning: assignment from incompatible pointer type
gsm-mms.c: In function ‘gn_mms_dec_miscstring’:
gsm-mms.c:315: warning: field width should have type ‘int’, but argument 3 has type ‘size_t’
gnokii.c: In function ‘shell’:
gnokii.c:1175: warning: assignment from incompatible pointer type
gnokii-file.c: In function ‘putfile’:
gnokii-file.c:423: warning: ‘error’ may be used uninitialized in this function
gnokii-utils.c: In function ‘readtext’:
gnokii-utils.c:151: warning: format ‘%d’ expects type ‘int’, but argument 4 has type ‘long unsigned int’

+++ common/gsm-mms.c
@@ -312,7 +312,7 @@ gn_error gn_mms_dec_miscstring(const unsigned char *source, size_t source_len, u
 			/* A null terminated string */
 			*decoded_len = strlen(&source[i]);
 		}
-		*dest_length = asprintf((char **) dest_buffer, "%*s", *decoded_len, &source[i]);
+		*dest_length = asprintf((char **) dest_buffer, "%*s", (int)*decoded_len, &source[i]);
 	}
 	(*decoded_len) += i;
